--- otptool
@@ -1,1 +1,0 @@
-./otptool.py
