+++ requirements.txt
@@ -3,3 +3,4 @@ jstyleson==0.0.2
 hexdump==3.3
 bitarray==1.0.1
 pycryptodome==3.9.7
+coverage==5.3.1
