--- socsec
@@ -1,1 +1,0 @@
-./socsec.py
