+++ tools/resocsec
